Fix memory leak test for SDPA op call (#168040)
Fixes #158330

Currently the test fails in the nightly because it keeps committing more work without waiting for previous work to finish and clean-up. So it is not testing the correct thing. Each iteration adds 2MB to the driver allocated memory as it holds on to the needed inputs, outputs and graph definition while the computation is in-flight.

Either waiting for about ~0.2s after each op call or calling device to synchronize allows the computation to wrap-up and the underlying memory to get released, displaying the expected result that the memory before and after the op call is the same.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/168040
Approved by: https://github.com/kulinseth