Add warning for clearing profiler events at the end of each cycle (#168066)
Fixes #148314.

This PR introduces a warning to clarify the behavior of the Profiler regarding event management. Specifically, it informs users that:
- The Profiler clears events at the end of each cycle.
- Only events from the current cycle will be reported by default.
- To retain events across cycles, users should set `acc_events=True`.
The warning is triggered only once when `self.profiler` is not `None` and `self.acc_events` is `False`, using the `warn_once` function.

This change aims to improve transparency and help users avoid confusion when analyzing profiling results.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/168066
Approved by: https://github.com/sraikund16