[xpu][fix] Refine memory pool logic when expandable segement enabled (#168956)
# Motivation
This is a bug in the interaction between the memory pool for XPUGraph and expandable segments.
When `unmap_block` is called, the allocator decreases `allocation_count` as expected:
(see lines 862–867)
https://github.com/pytorch/pytorch/blob/f3fb2d9df84f6759a43153fe7fc9e9cf308b92c5/c10/xpu/XPUCachingAllocator.cpp#L862-L867

However, when an expandable segment is created via `try_allocate_expandable_block`, we never increment `allocation_count`.
As a result, `allocation_count` can drop below its correct value after unmapping.

# Solution
This patch fixes the issue by ensuring `allocation_count` is incremented when creating a new expandable segment.

# Additional Context
PyTorch currently does not support using a custom allocator together with the expandable-segment feature in the memory pool. Therefore, we add an assertion to fail fast when this unsupported condition is detected.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/168956
Approved by: https://github.com/EikanWang