[xplat][caffe2] Fix -Wswitch-default issues (#169022)
Summary:
**Context:** https://fburl.com/switch-enum

----

Now that `-Wswitch-enum` is an error in all of `fbobjc`, making it such that all enum values need a case within a switch regardless of having a default case, we want to also enable `-Wswitch-default` as an error (ensuring we cover all values outside the given enum's values).  This will reduce SEVs as it will eliminate undefined behavior when a value outside the enum range is switched upon.

Test Plan: CI Pass

Differential Revision: D87817056

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169022
Approved by: https://github.com/shoumikhin