[DTensor] Efficient argmax and argmin (masked and unmasked) (#168983)
Fixes #167072
Contributes to #119930

This leverages the min/max operators to allow for linear reduction on sharded dimensions. The underlining code computes the global indices on the fly while reducing locally.  At the gathering of tensors because the min and max is tracked, the true indices will be returned.

CC: @H-Huang @awgu @wanchaol @fegin @fduwjj @wz337 @wconstab @d4l3k @pragupta @msaroufim @dcci
Pull Request resolved: https://github.com/pytorch/pytorch/pull/168983
Approved by: https://github.com/zpcore