[inductor] fix reading _TORCHINDUCTOR_PYOBJECT_TENSOR_DATA_PTR in module (#170012)
For an internal test, we're seeing the following with the old code.

Since `addr` is 0, but we didn't have a check on `iss.fail()`, the code will sefgault later when we run the compiled module.

```
_TORCHINDUCTOR_PYOBJECT_TENSOR_DATA_PTR 140719011151296
str_addr: [140719011151296]
parsed addr: 0
iss.fail(): 1, iss.eof(): 0, iss.bad(): 1
```

In this PR, we change to use a more robust parsing approach that avoids the C++ iostream issues, e.g. if some program corrupts locale. This PR pass the internal test.

We also add a check so the module load error out early if the parsing fails.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170012
Approved by: https://github.com/benjaminglass1, https://github.com/SherlockNoMad