[dynamo][hops] Ignore side effects for _reparameterize_module (#170251)
_reparameterize_module has two side effects that cancel each other out.
Even if we don't allow side effects in a region (will be used for
is_pure impl in future), we still want to ignore the side effects inside
this function. We already do this (check the impl) but we were not doing
this for hops.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170251
Approved by: https://github.com/zou3519