[Kineto Protobuf] update test_profiler to allow protobuf format data from log (#169868)
Summary:
In next diff, we create PerfettoTraceBuilder to generate protobuf format trace data and upload to manifold and log the url in the log. This test is checking if the logged trace format and the new protobuf trace is breaking it.

Let's remove the assertion

Test Plan: CI

Differential Revision: D88658203

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169868
Approved by: https://github.com/sraikund16