[xpu][feature] [1/6] Add trace support on XPU caching allocator (#168262)
# Motivation
This PR introduces the `TraceEntry` structure and the `recordHistory` functionality to enable trace support in the XPU caching allocator. At this stage, the implementation is XPU-specific; however, we plan to refactor and consolidate this logic with the existing CUDA implementation to maximize code reuse and consistency across backends. The design principle is to remain minimal and straightforward while laying the foundation for future enhancements.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/168262
Approved by: https://github.com/EikanWang, https://github.com/albanD