Fix cooperative reduction deadlocks due to too many CTAs (#170162)
Cooperative reductions currently use a hardcoded target number of CTAs of 64. This means that on systems with fewer than 64 SMs (e.g. Nvidia L4), we may have more CTAs than SMs. This is a problem because there are points where each CTA blocks waiting for all other CTAs computing the same reduction value to complete. If there are more CTAs than SMs, some of those CTAs may never get scheduled, so we deadlock.

This does increase the target number of CTAs on platforms that have more than 64 SMs. If this turns out to be bad, we can cap it back at 64 so we fix the correctness issue while maintaing the current behaviour for other GPUs.

Even with this change, I have some concerns about whether we can guarantee that this will not deadlock, especially if it is possible to have multiple kernels running on the GPU at the same time. For instance, is it possible to have two cooperative reductions scheduled simultaneously, such that only half of the CTAs in each reduction actually get to run?

Addresses some of the failures identified in #141915.

Fixes #169492

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170162
Approved by: https://github.com/jansel