[Dynamo][Guards]Add e2e user stack to guard debug info and ++ leaf guard (#169999)
Fix https://github.com/pytorch/pytorch/issues/167898

## Problem
Recompilation is expensive and it would be nice to reduce the number of such actions. The guard recompilations reason logs currently capture the last line of the stack and it doesn’t give enough information. By providing a full user stack trace, we would make finding the root cause of recompilation a lot easier and reduce the total amount of recompilations.

## Solution
An additional field is being implemented in LeafGuard to store the complete user stack, which will then be extracted via Guard debug information. This approach is expected to be successful because it isolates specific guards and only provides the necessary information when a guard fails and recompiles.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169999
Approved by: https://github.com/anijain2305