Make CachingHostAllocator work with memory pools. (#167507)
Both allocation to a cuda graph's private pool via stream capture and
allocation to a memory pool in non-stream-captured code are supported.

In the case of stream capture, we refuse to reuse a host memory block
as soon as record_event() is called on that block. This is to prevent
a stream-captured CUDA kernel from reading different contents from a
memory block than would be read if, counterfactually, this CUDA
kernels were running eagerly on a cuda stream.

See
https://github.com/pytorch/pytorch/pull/161583#issuecomment-3229885771
for elaboration.

This is lacking test cases for pagedable host memory copies. We must
make sure that record_event() does not fail in that case.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/167507
Approved by: https://github.com/eqy, https://github.com/eee4017, https://github.com/ngimel