[Inductor] Allow statically analyzing epilogues (#170001)
Differential Revision: D88786360

Now in the case of epilogue_fusion, we can have MultiTemplateBuffers that aren't necessarily returned from benchmark_epilogue_fusion due to the advent of Async-Pipelined Autotuning. In that case, we don't always want to default to benchmarking epilogues due to compile time overhead. We enforce some level of  safety in checking that the fusion compiles/has minimal register spillage and then take the fusion.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170001
Approved by: https://github.com/eellison