[DTensor] Hook up output tensor_meta to expand util (#170827)
Enforce tensor_meta is not none for new single-dim rules.

Allow tensor_meta to continue to be None for existing rules for now. We
should consider in the future asserting tensor_meta is required in
DTensorSpec, but for now we just try to limit the bleeding.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/170827
Approved by: https://github.com/dolpm
ghstack dependencies: #170615, #167677, #170359