[ez] log why fail to gen golden ref (#170843)
Sometimes when doing accuracy test with a large batch size, it fail due to the golden ref can not be generated. It would be helpful to log why we fail to generate the golden ref. In my case, it's due to OOM.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170843
Approved by: https://github.com/malfet, https://github.com/Skylion007